Remove LuaGlobals and move all functions into LuaEntity.cpp
